Destination S3 Data Lake: Extract generic iceberg stuff to toolkit (#53700)
