+++ airbyte-integrations/connectors/destination-s3-data-lake/src/main/kotlin/io/airbyte/integrations/destination/s3_data_lake/io/S3DataLakeUtil.kt
@@ -4,21 +4,13 @@
 
 package io.airbyte.integrations.destination.s3_data_lake.io
 
-import io.airbyte.cdk.load.command.Dedupe
 import io.airbyte.cdk.load.command.DestinationStream
-import io.airbyte.cdk.load.command.ImportType
 import io.airbyte.cdk.load.command.aws.AwsAssumeRoleCredentials
 import io.airbyte.cdk.load.command.iceberg.parquet.GlueCatalogConfiguration
 import io.airbyte.cdk.load.command.iceberg.parquet.IcebergCatalogConfiguration
 import io.airbyte.cdk.load.command.iceberg.parquet.NessieCatalogConfiguration
 import io.airbyte.cdk.load.command.iceberg.parquet.RestCatalogConfiguration
-import io.airbyte.cdk.load.data.MapperPipeline
-import io.airbyte.cdk.load.data.NullValue
-import io.airbyte.cdk.load.data.ObjectValue
-import io.airbyte.cdk.load.data.iceberg.parquet.toIcebergRecord
-import io.airbyte.cdk.load.data.iceberg.parquet.toIcebergSchema
-import io.airbyte.cdk.load.data.withAirbyteMeta
-import io.airbyte.cdk.load.message.DestinationRecordAirbyteValue
+import io.airbyte.cdk.load.toolkits.iceberg.parquet.io.IcebergUtil
 import io.airbyte.integrations.destination.s3_data_lake.ACCESS_KEY_ID
 import io.airbyte.integrations.destination.s3_data_lake.ASSUME_ROLE_ARN
 import io.airbyte.integrations.destination.s3_data_lake.ASSUME_ROLE_EXTERNAL_ID
@@ -29,187 +21,47 @@ import io.airbyte.integrations.destination.s3_data_lake.AWS_CREDENTIALS_MODE_STA
 import io.airbyte.integrations.destination.s3_data_lake.GlueCredentialsProvider
 import io.airbyte.integrations.destination.s3_data_lake.S3DataLakeConfiguration
 import io.airbyte.integrations.destination.s3_data_lake.SECRET_ACCESS_KEY
-import io.airbyte.integrations.destination.s3_data_lake.TableIdGenerator
 import io.github.oshai.kotlinlogging.KotlinLogging
 import jakarta.inject.Singleton
-import org.apache.hadoop.conf.Configuration
 import org.apache.iceberg.CatalogProperties
 import org.apache.iceberg.CatalogProperties.URI
 import org.apache.iceberg.CatalogUtil
 import org.apache.iceberg.CatalogUtil.ICEBERG_CATALOG_TYPE_GLUE
 import org.apache.iceberg.CatalogUtil.ICEBERG_CATALOG_TYPE_NESSIE
 import org.apache.iceberg.CatalogUtil.ICEBERG_CATALOG_TYPE_REST
-import org.apache.iceberg.FileFormat
-import org.apache.iceberg.Schema
-import org.apache.iceberg.SortOrder
-import org.apache.iceberg.Table
-import org.apache.iceberg.TableProperties.DEFAULT_FILE_FORMAT
 import org.apache.iceberg.aws.AwsClientProperties
 import org.apache.iceberg.aws.AwsProperties
 import org.apache.iceberg.aws.s3.S3FileIO
 import org.apache.iceberg.aws.s3.S3FileIOProperties
 import org.apache.iceberg.catalog.Catalog
-import org.apache.iceberg.catalog.SupportsNamespaces
-import org.apache.iceberg.data.Record
-import org.apache.iceberg.exceptions.AlreadyExistsException
 import org.projectnessie.client.NessieConfigConstants
-import software.amazon.awssdk.services.glue.model.ConcurrentModificationException
+
+private const val AWS_REGION = "aws.region"
 
 private val logger = KotlinLogging.logger {}
 
-const val AIRBYTE_CDC_DELETE_COLUMN = "_ab_cdc_deleted_at"
-const val EXTERNAL_ID = "AWS_ASSUME_ROLE_EXTERNAL_ID"
-const val AWS_ACCESS_KEY_ID = "AWS_ACCESS_KEY_ID"
-const val AWS_SECRET_ACCESS_KEY = "AWS_SECRET_ACCESS_KEY"
-private const val AWS_REGION = "aws.region"
-
 /**
  * Collection of Iceberg related utilities.
- * @param awsSystemCredentials is a temporary fix to allow us to run the integrations tests. This
+ * @param assumeRoleCredentials is a temporary fix to allow us to run the integrations tests. This
  * will be removed when we change all of this to use Micronaut
  */
 @Singleton
 class S3DataLakeUtil(
-    private val tableIdGenerator: TableIdGenerator,
+    private val icebergUtil: IcebergUtil,
     private val assumeRoleCredentials: AwsAssumeRoleCredentials?,
 ) {
-
-    internal class InvalidFormatException(message: String) : Exception(message)
-
-    private val generationIdRegex = Regex("""ab-generation-id-\d+-e""")
-
-    fun assertGenerationIdSuffixIsOfValidFormat(generationId: String) {
-        if (!generationIdRegex.matches(generationId)) {
-            throw InvalidFormatException(
-                "Invalid format: $generationId. Expected format is 'ab-generation-id-<number>-e'",
-            )
-        }
-    }
-
-    fun constructGenerationIdSuffix(stream: DestinationStream): String {
-        return constructGenerationIdSuffix(stream.generationId)
-    }
-
-    fun constructGenerationIdSuffix(generationId: Long): String {
-        if (generationId < 0) {
-            throw IllegalArgumentException(
-                "GenerationId must be non-negative. Provided: $generationId",
-            )
-        }
-        return "ab-generation-id-${generationId}-e"
-    }
-    /**
-     * Builds an Iceberg [Catalog].
-     *
-     * @param catalogName The name of the catalog.
-     * @param properties The map of catalog configuration properties.
-     * @return The configured Iceberg [Catalog].
-     */
-    fun createCatalog(catalogName: String, properties: Map<String, String>): Catalog {
-        return CatalogUtil.buildIcebergCatalog(catalogName, properties, Configuration())
-    }
-
-    /** Create the namespace if it doesn't already exist. */
-    fun createNamespace(streamDescriptor: DestinationStream.Descriptor, catalog: Catalog) {
-        val tableIdentifier = tableIdGenerator.toTableIdentifier(streamDescriptor)
-        synchronized(tableIdentifier.namespace()) {
-            if (
-                catalog is SupportsNamespaces &&
-                    !catalog.namespaceExists(tableIdentifier.namespace())
-            ) {
-                try {
-                    catalog.createNamespace(tableIdentifier.namespace())
-                    logger.info { "Created namespace '${tableIdentifier.namespace()}'." }
-                } catch (e: AlreadyExistsException) {
-                    // This exception occurs when multiple threads attempt to write to the same
-                    // namespace in parallel.
-                    // One thread may create the namespace successfully, causing the other threads
-                    // to encounter this exception
-                    // when they also try to create the namespace.
-                    logger.info {
-                        "Namespace '${tableIdentifier.namespace()}' was likely created by another thread during parallel operations."
-                    }
-                }
-            }
-        }
-    }
-
     fun createNamespaceWithGlueHandling(
         streamDescriptor: DestinationStream.Descriptor,
         catalog: Catalog
     ) {
         try {
-            createNamespace(streamDescriptor, catalog)
+            icebergUtil.createNamespace(streamDescriptor, catalog)
         } catch (e: ConcurrentModificationException) {
             // glue catalog throws its own special exception
             logger.info {
                 "Namespace '${streamDescriptor.namespace}' was likely created by another thread during parallel operations."
             }
         }
-    }
-
-    /**
-     * Builds (if necessary) an Iceberg [Table]. If the [Table] already exists, it is loaded from
-     * the [Catalog].
-     *
-     * Assumes the namespace already exists. Use [createNamespace] if this is not guaranteed.
-     *
-     * @param streamDescriptor The [DestinationStream.Descriptor] that contains the Airbyte stream's
-     * namespace and name.
-     * @param catalog The Iceberg [Catalog] that contains the [Table] or should contain it once
-     * created.
-     * @param schema The Iceberg [Schema] associated with the [Table].
-     * @param properties The [Table] configuration properties derived from the [Catalog].
-     * @return The Iceberg [Table], created if it does not yet exist.
-     */
-    fun createTable(
-        streamDescriptor: DestinationStream.Descriptor,
-        catalog: Catalog,
-        schema: Schema,
-        properties: Map<String, String>
-    ): Table {
-        val tableIdentifier = tableIdGenerator.toTableIdentifier(streamDescriptor)
-        return if (!catalog.tableExists(tableIdentifier)) {
-            logger.info { "Creating Iceberg table '$tableIdentifier'...." }
-            catalog
-                .buildTable(tableIdentifier, schema)
-                .withProperties(properties)
-                .withProperty(DEFAULT_FILE_FORMAT, FileFormat.PARQUET.name.lowercase())
-                .withSortOrder(getSortOrder(schema = schema))
-                .create()
-        } else {
-            logger.info { "Loading Iceberg table $tableIdentifier ..." }
-            catalog.loadTable(tableIdentifier)
-        }
-    }
-
-    /**
-     * Converts an Airbyte [DestinationRecordAirbyteValue] into an Iceberg [Record]. The converted
-     * record will be wrapped to include [Operation] information, which is used by the writer to
-     * determine how to write the data to the underlying Iceberg files.
-     *
-     * @param record The Airbyte [DestinationRecordAirbyteValue] record to be converted for writing
-     * by Iceberg.
-     * @param stream The Airbyte [DestinationStream] that contains information about the stream.
-     * @param tableSchema The Iceberg [Table] [Schema].
-     * @param pipeline The [MapperPipeline] used to convert the Airbyte record to an Iceberg record.
-     * @return An Iceberg [Record] representation of the Airbyte [DestinationRecordAirbyteValue].
-     */
-    fun toRecord(
-        record: DestinationRecordAirbyteValue,
-        stream: DestinationStream,
-        tableSchema: Schema,
-        pipeline: MapperPipeline
-    ): Record {
-        val dataMapped =
-            pipeline
-                .map(record.data, record.meta?.changes)
-                .withAirbyteMeta(stream, record.emittedAtMs, true)
-        // TODO figure out how to detect the actual operation value
-        return RecordWrapper(
-            delegate = dataMapped.toIcebergRecord(tableSchema),
-            operation = getOperation(record = record, importType = stream.importType)
-        )
     }
 
     /**
@@ -245,13 +97,13 @@ class S3DataLakeUtil(
         }
     }
 
+    // TODO this + nessie probably belong in base CDK toolkit
     private fun buildRestProperties(
         config: S3DataLakeConfiguration,
         catalogConfig: RestCatalogConfiguration,
         s3Properties: Map<String, String>,
         region: String
     ): Map<String, String> {
-        println(region)
         val awsAccessKeyId =
             requireNotNull(config.awsAccessKeyConfiguration.accessKeyId) {
                 "AWS Access Key ID is required for Rest configuration"
@@ -424,35 +276,4 @@ class S3DataLakeUtil(
 
         return properties
     }
-
-    fun toIcebergSchema(stream: DestinationStream, pipeline: MapperPipeline): Schema {
-        val primaryKeys =
-            when (stream.importType) {
-                is Dedupe -> (stream.importType as Dedupe).primaryKey
-                else -> emptyList()
-            }
-        return pipeline.finalSchema.withAirbyteMeta(true).toIcebergSchema(primaryKeys)
-    }
-
-    private fun getSortOrder(schema: Schema): SortOrder {
-        val builder = SortOrder.builderFor(schema)
-        schema.identifierFieldNames().forEach { builder.asc(it) }
-        return builder.build()
-    }
-
-    private fun getOperation(
-        record: DestinationRecordAirbyteValue,
-        importType: ImportType,
-    ): Operation =
-        if (
-            record.data is ObjectValue &&
-                (record.data as ObjectValue).values[AIRBYTE_CDC_DELETE_COLUMN] != null &&
-                (record.data as ObjectValue).values[AIRBYTE_CDC_DELETE_COLUMN] !is NullValue
-        ) {
-            Operation.DELETE
-        } else if (importType is Dedupe) {
-            Operation.UPDATE
-        } else {
-            Operation.INSERT
-        }
 }
